+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Add more logs to improve debugging.
+- Set `EnableKeepAlive=false` to avoid deadlock.
+
 ## [0.3.0] - 2024-07-24
 
 ### Changed